requirements.txt
PyPDF2==3.0.1
PyMuPDF
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
tiktoken
langchain==0.0.316
openai==0.28.1
streamlit-extras
altair
# pip install --upgrade setuptools wheel
# python 3.10.7

